requirements.txt226.txt
streamlit
google-api-python-client
google-generativeai
pandas